Refactors to add api folder
