+++ makefile
@@ -6,11 +6,11 @@ BIN_DIR?=.
 .PHONY: build
 build:
 	@mkdir -p $(BUILD)/$(BIN_DIR)
-	go build -o $(BUILD)/$(BIN_DIR)/books-api main.go book.go errors.go
+	go build -o $(BUILD)/$(BIN_DIR)/books-api main.go
 
 .PHONY: debug
 debug: build
-	HUMAN_LOG=1 go run -race main.go book.go errors.go
+	HUMAN_LOG=1 go run -race main.go
 
 .PHONY: test
 test:
